remodel audio stream
Completely changed the file. Made the audio to be a callback. Took only
the first bin as a stream of scalars to feed into an OPF model. This
opposed to taking the whole FFT vector and feeding directly into  a
temporal pooler.  Added anomaly and likelihood computations, and
plotted them.